add traction details class + part mappings for it

+++ godot/Cars/Sim/CalcWheelTraction.cs
@@ -15,7 +15,7 @@ public class CalcWheelTraction {
     /// _D = radians-ish, traction drop off after peak
     /// g(x)=If(x<_S, _T*((x)/(_S)), If(x>_S+_L, _T+(_S+_L-x) _D, _T))
     /// </summary>
-    public static float Calc(float slip, float s, float t, float l, float d) {
+    public static double Calc(double slip, double s, double t, double l, double d) {
         // because i am bad at math negative values go through as positive
         if (slip < 0) {
             return Mathf.Sign(slip) * Calc(Math.Abs(slip), s, t, l, d);
